requirements.txt
annotated-types==0.6.0
anyio==4.3.0
ecdsa==0.18.0
fastapi==0.110.0
idna==3.6
passlib==1.7.4
pyasn1==0.5.1
pydantic==2.6.3
pydantic_core==2.16.3
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.36.3
typing_extensions==4.10.0
